inital api attempt +organised some files
